[!!!][TASK] Don't respect sub-categories of user's categories
In commit 5dbc3a166d9bd50085d28cdfb42d7d5483fc0746 of EXT:news, the
AccessControlService was changed to respect the sub-categories of
the user's categories. Therefore, if a user has permissions to a
parent category, he also gets access to all sub-categories.

This is an unwanted behaviour in this context. Therefore, the
AccessControlService is overridden. Because its methods are static,
the classes using the service must also be overridden.